# api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "URL TRIM API"
  },
  "paths": {
    "/api/TrimUrl": {
      "post": {
        "tags": [
          "TrimUrl"
        ],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string",
                "nullable": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TrimUriModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrimUriModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrimUriModel"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "TrimUrl"
        ],
        "parameters": [
          {
            "name": "index",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
			  "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TrimUriModelListPageModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrimUriModelListPageModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrimUriModelListPageModel"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TrimUriModel": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "nullable": false,
            "readOnly": true
          },
          "trimmed": {
            "type": "string",
            "nullable": false,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "TrimUriModelListPageModel": {
        "type": "object",
        "properties": {
          "index": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "count": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrimUriModel"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}